test_wraptree: split the source and test code export
The tests are in the same directory structure as the source code. However, it
is necessary to give WRAPTREE_TESTS_ROOT its own task and location since
WRAPTREE_ROOT is a local product on a machine (any platform) which (cross)
compiled the executable. The tests need to run on a Window machine.