ape: Shut down CC with SIGTERM instead of SIGINT
While SIGINT works in my developer setup, it doesn't seem to work when
the CC is running in the background. SIGTERM does work and coverage
information is saved in that case, so we can use that.

I recall that earlier SIGTERM resulted in a loss of the coverage
information, but perhaps that was due to 'twist' setting up different
signal handlers.