poetry: Use 'install' instead of 'update' to prepare virtualenv
The problem with 'update' is that while it installs dependencies,
it does not force installation of the package itself, which is
a must now that SoftFab uses metadata to retrieve its version.

Running 'update' was necessary when Poetry would use one virtualenv
per Python version, but recent Poetry releases also include a hash
of the source path in the virtualenv name. So we now get a fresh
env for every job, which means updating is unnecessary and even
undesired.